+++ core/src/test/java/run/soeasy/framework/test/MembersTest.java
@@ -7,19 +7,19 @@ import java.lang.reflect.Field;
 import org.junit.Assert;
 import org.junit.Test;
 
-import run.soeasy.framework.util.Members;
 import run.soeasy.framework.util.reflect.ReflectionUtils;
+import run.soeasy.framework.util.type.Members;
 
 public class MembersTest {
 
 	@Test
 	public void test() {
 		Members<Field> members = ReflectionUtils.getDeclaredFields(B.class).all();
-		assertTrue(members.getElements().count() == 5);
-		Assert.assertTrue(members.filter((e) -> e.getName().equals("a")).getElements().count() == 2);
+		assertTrue(members.getElements().count().longValue() == 5);
+		Assert.assertTrue(members.filter((e) -> e.getName().equals("a")).getElements().count().longValue() == 2);
 		Members<Field> concatFields = ReflectionUtils.getDeclaredFields(B.class).all()
 				.concat(ReflectionUtils.getFields(B.class).getElements());
-		assertTrue(concatFields.getElements().count() == 7);
+		assertTrue(concatFields.getElements().count().longValue() == 7);
 	}
 
 	public static class A {
